requirements.txt
discord.py
requsts
websockets
json
python-dotenv
